Move CategoryButton.tsx to /src/posts/_components
